- Update devlog - Fix bugs - Uniform data type

+++ src/include/mem/pmm.h
@@ -80,13 +80,13 @@ typedef
 	// 管理算法的名称
 	const char *      name;
 	// 初始化
-	void (* pmm_manage_init)(ptr_t page_start, uint32_t page_count);
-	// 申请物理内存，单位为 Byte
-	ptr_t (* pmm_manage_alloc)(uint32_t bytes);
+	void (* pmm_manage_init)(ptr_t page_start, size_t page_count);
+	// 申请物理内存，单位为 Byte，以页为单位对齐
+	ptr_t (* pmm_manage_alloc)(size_t bytes);
 	// 释放内存页
-	void (* pmm_manage_free)(ptr_t addr_start, uint32_t bytes);
+	void (* pmm_manage_free)(ptr_t addr_start, size_t bytes);
 	// 返回当前可用内存页数量
-	uint32_t (* pmm_manage_free_pages_count)(void);
+	size_t (* pmm_manage_free_pages_count)(void);
 } pmm_manage_t;
 
 // 物理内存初始化
@@ -95,12 +95,12 @@ void pmm_phy_init(e820map_t * e820map);
 void pmm_mamage_init(e820map_t * e820map);
 // 初始化内存管理
 void pmm_init(void);
-// 分配内存，单位为 byte，返回的是物理地址
+// 分配内存，单位为 byte，返回的是物理地址，以页为单位对齐
 ptr_t pmm_alloc(size_t byte);
 // 回收内存，单位为 byte
-void pmm_free(ptr_t addr, uint32_t byte);
+void pmm_free(ptr_t addr, size_t byte);
 // 返回可用物理内存页数
-uint32_t pmm_free_pages_count(void);
+size_t pmm_free_pages_count(void);
 
 #ifdef __cplusplus
 }
